correctly recognize terminated jobs
addresses #17
