+++ jobmanager/io/io.py
@@ -172,10 +172,10 @@ def read_outfile(outfile_path, short_ouput=False, long_output=True):
             if is_finished[0] == 'Job' and is_finished[1] == 'finished:':
                 finished = True
 
-        is_finished = output.wordgrab(['processing'], 'whole_line', last_line=True)[0]
-        if is_finished:  # for hydrogen optimization
-            if is_finished[0] == 'Total' and is_finished[1] == 'processing':
-                finished = True
+        # is_finished = output.wordgrab(['processing'], 'whole_line', last_line=True)[0]
+        # if is_finished:  # for hydrogen optimization
+        #     if is_finished[0] == 'Total' and is_finished[1] == 'processing':
+        #         finished = True
 
         #terminated
         is_terminated = output.wordgrab(['terminated:'],'whole_line',last_line=True)[0]
